requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.7
GeoAlchemy2==0.17.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
